+++ Classes/Score.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SudokuApp
+{
+    class Score
+    {
+        public int Points { get; set; }
+        public int Min { get; set; }
+        public int Sec { get; set; }
+        public bool Visible { get; set; }
+    }
+}
